requirements.txt
# Common dependencies
aiohttp
fastapi[all]>=0.104.0
uvicorn>=0.24.0
httpx
asyncpg
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
psutil
pandas
openai>=1.3.0
requests>=2.31.0
tenacity>=8.0.0
gunicorn>=21.2.0  # Add gunicorn
cachetools>=5.3.0  # Add this line
torch
sqlalchemy

# Database dependencies
psycopg2-binary>=2.9.9

# Kafka and retry dependencies
kafka-python[sasl]>=2.0.2
backoff>=2.2.1